nfdiv-1183: move internal constants back to classes
