+++ src/test/java/uk/gov/hmcts/divorce/caseworker/service/notification/NoticeOfProceedingsNotificationTest.java
@@ -18,6 +18,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
+import static uk.gov.hmcts.divorce.caseworker.service.notification.NoticeOfProceedingsNotification.CASE_ID;
+import static uk.gov.hmcts.divorce.caseworker.service.notification.NoticeOfProceedingsNotification.SOLICITOR_ORGANISATION;
 import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.ENGLISH;
 import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.WELSH;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.APPLICANT_NOTICE_OF_PROCEEDINGS;
@@ -26,10 +28,8 @@ import static uk.gov.hmcts.divorce.notification.EmailTemplateName.RESPONDENT_SOL
 import static uk.gov.hmcts.divorce.notification.FormatUtil.formatId;
 import static uk.gov.hmcts.divorce.notification.NotificationConstants.APPLICANT_NAME;
 import static uk.gov.hmcts.divorce.notification.NotificationConstants.APPLICATION_REFERENCE;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.CASE_ID;
 import static uk.gov.hmcts.divorce.notification.NotificationConstants.RESPONDENT_NAME;
 import static uk.gov.hmcts.divorce.notification.NotificationConstants.SOLICITOR_NAME;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.SOLICITOR_ORGANISATION;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.APPLICANT_2_FIRST_NAME;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_FIRST_NAME;
